requirements.txt
flask
pandas
jinja2
python-dotenv
openpyxl
langdetect